+++ framework-system/framework-management/src/test/java/uk/gov/justice/services/management/shuttering/handler/ShutteringSystemCommandHandlerTest.java
@@ -3,13 +3,19 @@ package uk.gov.justice.services.management.shuttering.handler;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyZeroInteractions;
 import static org.mockito.Mockito.when;
+import static uk.gov.justice.services.management.shuttering.observers.shuttering.ContextShutteredState.CONTEXT_SHUTTERED;
+import static uk.gov.justice.services.management.shuttering.observers.shuttering.ContextShutteredState.CONTEXT_UNSHUTTERED;
+import static uk.gov.justice.services.management.shuttering.observers.shuttering.ContextShutteredState.SHUTTERING_STARTED;
+import static uk.gov.justice.services.management.shuttering.observers.shuttering.ContextShutteredState.UNSHUTTERING_STARTED;
 
 import uk.gov.justice.services.common.util.UtcClock;
 import uk.gov.justice.services.jmx.api.command.ShutterSystemCommand;
 import uk.gov.justice.services.jmx.api.command.UnshutterSystemCommand;
 import uk.gov.justice.services.management.shuttering.events.ShutteringRequestedEvent;
 import uk.gov.justice.services.management.shuttering.events.UnshutteringRequestedEvent;
+import uk.gov.justice.services.management.shuttering.observers.ContextShutteredStateObserver;
 
 import java.time.ZonedDateTime;
 
@@ -22,6 +28,7 @@ import org.mockito.Captor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
+import org.slf4j.Logger;
 
 @RunWith(MockitoJUnitRunner.class)
 public class ShutteringSystemCommandHandlerTest {
@@ -34,6 +41,12 @@ public class ShutteringSystemCommandHandlerTest {
 
     @Mock
     private Event<UnshutteringRequestedEvent> unshutteringRequestedEventFirer;
+
+    @Mock
+    private ContextShutteredStateObserver contextShutteredStateObserver;
+
+    @Mock
+    private Logger logger;
 
     @InjectMocks
     private ShutteringSystemCommandHandler shutteringSystemCommandHandler;
@@ -50,6 +63,7 @@ public class ShutteringSystemCommandHandlerTest {
         final ZonedDateTime now = new UtcClock().now();
         final ShutterSystemCommand shutterSystemCommand = new ShutterSystemCommand();
 
+        when(contextShutteredStateObserver.getShutteredState()).thenReturn(CONTEXT_UNSHUTTERED);
         when(clock.now()).thenReturn(now);
 
         shutteringSystemCommandHandler.onShutterRequested(shutterSystemCommand);
@@ -68,6 +82,7 @@ public class ShutteringSystemCommandHandlerTest {
         final ZonedDateTime now = new UtcClock().now();
         final UnshutterSystemCommand unshutterSystemCommand = new UnshutterSystemCommand();
 
+        when(contextShutteredStateObserver.getShutteredState()).thenReturn(CONTEXT_SHUTTERED);
         when(clock.now()).thenReturn(now);
 
         shutteringSystemCommandHandler.onUnshutterRequested(unshutterSystemCommand);
@@ -78,5 +93,71 @@ public class ShutteringSystemCommandHandlerTest {
 
         assertThat(unshutteringRequestedEvent.getTarget(), is(unshutterSystemCommand));
         assertThat(unshutteringRequestedEvent.getUnshutteringRequestedAt(), is(now));
+    }
+
+    @Test
+    public void shouldIgnoreShutteringRequestIfTheShutteredStateIsContextShuttered() throws Exception {
+
+        when(contextShutteredStateObserver.getShutteredState()).thenReturn(CONTEXT_SHUTTERED);
+
+        shutteringSystemCommandHandler.onShutterRequested(new ShutterSystemCommand());
+
+        verifyZeroInteractions(shutteringRequestedEventFirer);
+        verify(logger).info("Ignoring command 'SHUTTER'. Context shuttered state is 'CONTEXT_SHUTTERED'");
+    }
+
+    @Test
+    public void shouldIgnoreShutteringRequestIfTheShutteredStateIsShutteringStarted() throws Exception {
+
+        when(contextShutteredStateObserver.getShutteredState()).thenReturn(SHUTTERING_STARTED);
+
+        shutteringSystemCommandHandler.onShutterRequested(new ShutterSystemCommand());
+
+        verifyZeroInteractions(shutteringRequestedEventFirer);
+        verify(logger).info("Ignoring command 'SHUTTER'. Context shuttered state is 'SHUTTERING_STARTED'");
+    }
+
+    @Test
+    public void shouldIgnoreShutteringRequestIfTheShutteredStateIsUnshutteringStarted() throws Exception {
+
+        when(contextShutteredStateObserver.getShutteredState()).thenReturn(UNSHUTTERING_STARTED);
+
+        shutteringSystemCommandHandler.onShutterRequested(new ShutterSystemCommand());
+
+        verifyZeroInteractions(shutteringRequestedEventFirer);
+        verify(logger).info("Ignoring command 'SHUTTER'. Context shuttered state is 'UNSHUTTERING_STARTED'");
+    }
+
+    @Test
+    public void shouldIgnoreUnshutteringRequestIfTheShutteredStateIsContextUnshuttered() throws Exception {
+
+        when(contextShutteredStateObserver.getShutteredState()).thenReturn(CONTEXT_UNSHUTTERED);
+
+        shutteringSystemCommandHandler.onUnshutterRequested(new UnshutterSystemCommand());
+
+        verifyZeroInteractions(shutteringRequestedEventFirer);
+        verify(logger).info("Ignoring command 'UNSHUTTER'. Context shuttered state is 'CONTEXT_UNSHUTTERED'");
+    }
+
+    @Test
+    public void shouldIgnoreUnshutteringRequestIfTheShutteredStateIsShutteringStarted() throws Exception {
+
+        when(contextShutteredStateObserver.getShutteredState()).thenReturn(SHUTTERING_STARTED);
+
+        shutteringSystemCommandHandler.onUnshutterRequested(new UnshutterSystemCommand());
+
+        verifyZeroInteractions(shutteringRequestedEventFirer);
+        verify(logger).info("Ignoring command 'UNSHUTTER'. Context shuttered state is 'SHUTTERING_STARTED'");
+    }
+
+    @Test
+    public void shouldIgnoreUnshutteringRequestIfTheShutteredStateIsUnshutteringStarted() throws Exception {
+
+        when(contextShutteredStateObserver.getShutteredState()).thenReturn(UNSHUTTERING_STARTED);
+
+        shutteringSystemCommandHandler.onUnshutterRequested(new UnshutterSystemCommand());
+
+        verifyZeroInteractions(shutteringRequestedEventFirer);
+        verify(logger).info("Ignoring command 'UNSHUTTER'. Context shuttered state is 'UNSHUTTERING_STARTED'");
     }
 }